implement ability to backup state in case it corrupts
