+++ Cargo.toml
@@ -44,7 +44,6 @@ log.workspace = true
 rust_decimal.workspace = true
 uuid.workspace = true
 fern = "0.6.2"
-dirs-next = "2.0.0"
 
 exchanges = { version = "0.1.0", path = "exchanges" }
 data = { version = "0.1.0", path = "data" }
