+++ data/Cargo.toml
@@ -11,5 +11,7 @@ chrono.workspace = true
 ordered-float.workspace = true
 log.workspace = true
 thiserror.workspace = true
+regex.workspace = true
+dirs-next = "2.0.0"
 
 exchanges = { version = "0.1.0", path = "../exchanges" }
